refactor: clothes 패키지 application Command, Query로 분리
