+++ src/main/java/dandi/dandi/clothes/application/port/in/ClothesQueryServicePort.java
@@ -1,21 +1,16 @@
 package dandi.dandi.clothes.application.port.in;
 
-
 import java.time.LocalDate;
 import java.util.Set;
 import org.springframework.data.domain.Pageable;
 
-public interface ClothesUseCase {
-
-    void registerClothes(Long memberId, ClothesRegisterCommand clothesRegisterCommand);
+public interface ClothesQueryServicePort {
 
     ClothesDetailResponse getSingleClothesDetails(Long memberId, Long clothesId);
 
     ClothesResponses getClothes(Long memberId, String category, Set<String> season, Pageable pageable);
 
     CategorySeasonsResponses getCategoriesAndSeasons(Long memberId);
-
-    void deleteClothes(Long memberId, Long clothesId);
 
     ClothesResponses getTodayClothes(Long memberId, LocalDate today, Pageable pageable);
 }
