+++ src/main/java/dandi/dandi/clothes/application/port/in/ClothesUseCasePort.java
@@ -1,0 +1,9 @@
+package dandi.dandi.clothes.application.port.in;
+
+
+public interface ClothesUseCasePort {
+
+    void registerClothes(Long memberId, ClothesRegisterCommand clothesRegisterCommand);
+
+    void deleteClothes(Long memberId, Long clothesId);
+}
