+++ src/main/java/dandi/dandi/clothes/application/service/ClothesQueryServiceAdapter.java
@@ -3,10 +3,9 @@ package dandi.dandi.clothes.application.service;
 import dandi.dandi.clothes.application.port.in.CategorySeasonsResponse;
 import dandi.dandi.clothes.application.port.in.CategorySeasonsResponses;
 import dandi.dandi.clothes.application.port.in.ClothesDetailResponse;
-import dandi.dandi.clothes.application.port.in.ClothesRegisterCommand;
+import dandi.dandi.clothes.application.port.in.ClothesQueryServicePort;
 import dandi.dandi.clothes.application.port.in.ClothesResponse;
 import dandi.dandi.clothes.application.port.in.ClothesResponses;
-import dandi.dandi.clothes.application.port.in.ClothesUseCase;
 import dandi.dandi.clothes.application.port.out.persistence.CategorySeasonProjection;
 import dandi.dandi.clothes.application.port.out.persistence.ClothesPersistencePort;
 import dandi.dandi.clothes.domain.Category;
@@ -30,37 +29,21 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class ClothesService implements ClothesUseCase {
+@Transactional(readOnly = true)
+public class ClothesQueryServiceAdapter implements ClothesQueryServicePort {
 
-    private static final int CLOTHES_IMAGE_URL_INDEX = 1;
     private static final String ALL = "ALL";
 
     private final ClothesPersistencePort clothesPersistencePort;
-    private final ClothesImageService clothesImageService;
     private final String imageAccessUrl;
 
-    public ClothesService(ClothesPersistencePort clothesPersistencePort, ClothesImageService clothesImageService,
-                          @Value("${cloud.aws.cloud-front.uri}") String imageAccessUrl) {
+    public ClothesQueryServiceAdapter(ClothesPersistencePort clothesPersistencePort,
+                                      @Value("${cloud.aws.cloud-front.uri}") String imageAccessUrl) {
         this.clothesPersistencePort = clothesPersistencePort;
-        this.clothesImageService = clothesImageService;
         this.imageAccessUrl = imageAccessUrl;
     }
 
     @Override
-    @Transactional
-    public void registerClothes(Long memberId, ClothesRegisterCommand clothesRegisterCommand) {
-        String clothesImageUrl = removeImageAccessUrl(clothesRegisterCommand.getClothesImageUrl());
-        Clothes clothes = Clothes.initial(memberId, clothesRegisterCommand.getCategory(),
-                clothesRegisterCommand.getSeasons(), clothesImageUrl);
-        clothesPersistencePort.save(clothes);
-    }
-
-    private String removeImageAccessUrl(String clothesImageUrl) {
-        return clothesImageUrl.split(imageAccessUrl)[CLOTHES_IMAGE_URL_INDEX];
-    }
-
-    @Override
-    @Transactional(readOnly = true)
     public CategorySeasonsResponses getCategoriesAndSeasons(Long memberId) {
         Map<Category, EnumSet<Season>> categoriesAndSeasons = new EnumMap<>(Category.class);
         for (CategorySeasonProjection categorySeason :
@@ -99,7 +82,6 @@ public class ClothesService implements ClothesUseCase {
     }
 
     @Override
-    @Transactional(readOnly = true)
     public ClothesDetailResponse getSingleClothesDetails(Long memberId, Long clothesId) {
         Clothes clothes = clothesPersistencePort.findById(clothesId)
                 .orElseThrow(NotFoundException::clothes);
@@ -137,23 +119,6 @@ public class ClothesService implements ClothesUseCase {
     }
 
     @Override
-    @Transactional
-    public void deleteClothes(Long memberId, Long clothesId) {
-        Clothes clothes = clothesPersistencePort.findById(clothesId)
-                .orElseThrow(NotFoundException::clothes);
-        validateOwner(clothes, memberId);
-        clothesPersistencePort.deleteById(clothesId);
-        clothesImageService.deleteClothesImage(clothes);
-    }
-
-    private void validateOwner(Clothes clothes, Long memberId) {
-        if (!clothes.isOwnedBy(memberId)) {
-            throw ForbiddenException.clothesDeletion();
-        }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
     public ClothesResponses getTodayClothes(Long memberId, LocalDate today, Pageable pageable) {
         Month month = Month.fromDate(today);
         Set<Season> seasons = month.getSeasons();
